Phase 1 of making the tosa-partition pass into a generic utility.
Simplify isConstantZero() and make it more general.

Restructure partition walk to collect anchors then process them
back-to-front, and drop transpose special case.

Update rocmlir tests that are perturbed by the partitioning rearrangement.

Rearranged the code to make a stub TosaPartitionPass with all the action
in a tosa-free Partition class.  Adjusted one test for different
behavior with tosa::transpose.

Create lambdas outside of arglist to avoid stack-use-after-free.

Save the Type, not the Value, of params, and search them last.
Amazingly, it seems to work.

This is a state that incorporates most of Simon's pull request, but not the
traceInputs() change to the partitioner.  It does pass all the Dialect/Tosa tests.
In sum:  generic partitioning, Tosa implementation, some of Simon's changes.

Clean up erroneous patches and compilation and test problems after rebase.

Tweak traceInputs() and use it on trailing-op operands like Simon does,
and some cleanup.

Put trailingOps' traced inputs in leadingOps to avoid perturbing
trailingOps.  Adjust tests accordingly.

First cut at separating the partitioner into a separate utility, used by tosa-partition.

Rename partitioner utility to outliner.

Changes from the review.
