+++ mlir/test/xmir/mobilenetv1.mid.mlir
@@ -1,14 +1,10 @@
 // RUN: rocmlir-driver -host-pipeline partition,highlevel -targets gfx908 %s | FileCheck %s
 
 module {
-
-// CHECK:  func.func private @mobilenetv1__part_4(%arg0: memref<1x224x224x3xf32> {func.read_access}, %arg1: memref<32x3x3x3xf32> {func.read_access}, %arg2: memref<1x112x112x32xf32> {func.write_access}) {
-
-// CHECK: func.func @mobilenetv1(%arg0: memref<1x224x224x3xf32>, %arg1: memref<32x3x3x3xf32>, %arg2: memref<3x3x32x1xf32>, %arg3: memref<64x1x1x32xf32>, %arg4: memref<3x3x64x1xf32>, %arg5: memref<128x1x1x64xf32>, %arg6: memref<1x56x56x128xf32>) {
-// CHECK:   %[[TOKEN0:.*]] = async.launch @mobilenetv1__part_4 (%arg0, %arg1, %{{.*}}) : (memref<1x224x224x3xf32>, memref<32x3x3x3xf32>, memref<1x112x112x32xf32>)
-
-// CHECK-DIS:   %[[TOKEN1:.*]] = async.launch @mobilenetv1__part_3 (%{{.*}}, %arg2, %{{.*}}) : (memref<1x112x112x32xf32>, memref<3x3x32x1xf32>, memref<12544x32xf32>)
-
+// CHECK:  func.func private @mobilenetv1__part_1(%arg0: memref<1x112x112x64xf32> {func.read_access}, %arg1: memref<3x3x64x1xf32> {func.read_access}, %arg2: memref<1x56x56x64xf32> {func.write_access}) {
+// CHECK:  func.func @mobilenetv1(%arg0: memref<1x224x224x3xf32>, %arg1: memref<32x3x3x3xf32>, %arg2: memref<3x3x32x1xf32>, %arg3: memref<64x1x1x32xf32>, %arg4: memref<3x3x64x1xf32>, %arg5: memref<128x1x1x64xf32>, %arg6: memref<1x56x56x128xf32>) {
+// CHECK:  %[[TOKEN0:.*]] = async.launch @mobilenetv1__part_4 (%arg0, %arg1, %{{.*}}) : (memref<1x224x224x3xf32>, memref<32x3x3x3xf32>, memref<1x112x112x32xf32>)
+// CHECK-DIS:   %[[TOKEN1:.*]] = async.launch @mobilenetv1__part_3 %[[TOKEN0:.*]] (%{{.*}}, %arg2, %{{.*}}) : (memref<1x112x112x32xf32>, memref<3x3x32x1xf32>, memref<1x112x112x32xf32>)
 // CHECK:   async.await %token_{{.*}} : !async.token
 
   func.func @mobilenetv1(%input_image: tensor<1x224x224x3xf32>, %f0: tensor<32x3x3x3xf32>, %f1: tensor<3x3x32x1xf32>, %f2: tensor<64x1x1x32xf32>, %f3: tensor<3x3x64x1xf32>, %f4: tensor<128x1x1x64xf32>) -> tensor<1x56x56x128xf32> {
