+++ mlir/test/xmir/mobilenetv1.small.mlir
@@ -7,7 +7,7 @@ module {
 // CHECK:  func.func @mobilenetv1(%arg0: tensor<1x112x112x32xf32>, %arg1: tensor<32x3x3x3xf32>, %arg2: tensor<3x3x32x1xf32>, %arg3: tensor<64x1x1x32xf32>) -> tensor<1x112x112x64xf32> {
 // CHECK:    %[[T0:.*]], %[[RES0:.*]] = async.launch @mobilenetv1__part_1 (%arg0, %arg2) : (tensor<1x112x112x32xf32>, tensor<3x3x32x1xf32>) -> tensor<1x112x112x32xf32>
 // CHECK:    %[[T1:.*]], %[[RES1:.*]] = async.launch @mobilenetv1__part_0
-// CHECK-SAME:[%[[T0]]] 
+// CHECK-SAME:[%[[T0]]]
 // CHECK-SAME:(%[[RES0]]
 // CHECK-SAME: %arg3) : (tensor<1x112x112x32xf32>, tensor<64x1x1x32xf32>) -> tensor<1x112x112x64xf32>
