+++ mlir/test/xmir/resnet50_blk.async.mlir
@@ -24,4 +24,3 @@ module {
     return %4 : tensor<1x32x32x64xf32>
   }
 }
-
